backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.34.127
botocore==1.34.127
certifi==2024.6.2
charset-normalizer==3.3.2
cohere==5.5.7
fastavro==1.9.4
filelock==3.15.1
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
jmespath==1.0.1
packaging==24.1
parameterized==0.9.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
tokenizers==0.15.2
tqdm==4.66.4
types-requests==2.32.0.20240602
typing_extensions==4.12.2
urllib3==2.2.1
